Create main.py
Welcome to the tip calculator. It's going to ask you for an input for how much the total bill came to. Let's say that we and a couple of friends went out for lunch and it came to 124.56 dollars.

And then it'll ask you, what percentage tip would you like to give? So let's say we want to give a 12 percent tip. We enter 12 and hit enter. And then finally it asks how many people are splitting this bill? Let's say there's seven of us and then hit enter.

Finally, it's going to tell us that each person should pay about $19.93.